backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.1

# ── NLP & Embeddings ──────────────────────────────────────────────────────────
spacy==3.7.5
# After install, run: python -m spacy download en_core_web_sm
sentence-transformers==2.7.0
torch==2.3.0

# ── Resume Parsing ────────────────────────────────────────────────────────────
pdfplumber==0.11.0
python-docx==1.1.0

# ── OneDrive / Microsoft Graph ────────────────────────────────────────────────
msal==1.28.0
requests==2.32.3
httpx==0.27.0

# ── Utilities ─────────────────────────────────────────────────────────────────
loguru==0.7.2
aiofiles==23.2.1
python-dotenv==1.0.1
numpy==1.26.4
